Cleanup: remove custom type entry.
In reality, we want to support custom types (see bug 765226 [1]), but the
code for that was basically removed in commit 331f5ba. For now, we remove
that code so we can do a better job at cleaning up things.

Note: in the future, we should probably use Gtk.Combobox.with_entry() instead.

[1] https://bugzilla.gnome.org/show_bug.cgi?id=765226<Paste>